[feature 162171018] code for the user-password-reset
