+++ index.js
@@ -3,11 +3,13 @@ import morgan from 'morgan';
 import bodyParser from 'body-parser';
 import nodeLogger from 'logger';
 import validator from 'express-validator';
-import { auth, profiles, user } from './routes';
+import {
+  auth, profiles, user, passwords
+} from './routes';
 
 const logger = nodeLogger.createLogger();
 
-const PORT = process.env.PORT || 4000;
+const PORT = process.env.PORT || 8001;
 
 const app = express();
 
@@ -15,7 +17,8 @@ app.use(morgan('dev'));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(validator());
-app.use('/api/v1/forgotpassword', user);
+app.use('/api/v1/passwords', passwords);
+app.use('/api/v1/users', user);
 app.use('/api/v1', auth);
 app.use('/api/v1/profiles', profiles);
