+++ routes/index.js
@@ -1,5 +1,8 @@
 import auth from './auth';
 import user from './users';
 import profiles from './profiles';
+import passwords from './passwords';
 
-export { auth, profiles, user };
+export {
+  auth, profiles, user, passwords
+};
